Use BaseMixin
